Add modality test to automated suite
